Fix rsync directories
Rsync uses a trailing slash on directories to mean "the contents of this
directory". That is, `rsync a b` and `rsync a/ b/` are subtly different:

If directory `a` contained a file `thing`, the first puts a inside b,
so you end up with `b/a/thing`. The second puts the contents of a inside
b, so you end up with `b/thing`.

We want the second type when dealing with the local repository backups.