Download phantomjs in to our /opt/dist packaged dependencies
I'm setting the var `phantomjs_download_location` to `/tmp` in the
phantomjs role so that it can always download even when the dependencies
role isn't being used. The dependencies role then creates `/opt/dist/tars`
and overrides the download location, forcing the phantomjs role to use it's
location.

It's a bit ugly but means that one role doesn't rely on stuff another has
created, which is cleaner.