Only package git repos on jenkins
We'll use the `package_dependencies` host for jenkins and the
`package_lighthouse` host for lighthouse. Since jenkins will be a local
connection we won't need to collect the dependencies (since they are
already in our desired location) but we do want to create the yum repo,
so we guarantee that the no-internet environment can install from file.