Updated README
Reordered a few things and linked to some external guides (mainly to do
with creating or updating keys). Fleshed out a few bits, too, and added
some more detail where necessary. The README now explains the role of
Terraform in a little more detail and how to use it. It also contains
more info about `servers.yml` and how `uWSGI` and `nginx` work together
in production. Jenkins was fleshed out and restructured a little, too.