Copyright notices for builder.
I added copyright using a slightly modified version of the script in the lighthouse
repository. The paths that it doesn't look in for adding copyright are mainly the
third-party ansible roles by guy geerling and of course rob (octarine). Also had to exclude
the secrets repo. Fairly sure all the things that are commentable have been commented here.

I included the script too, so it can be used again later.