Pin ansible to 2.1.1.0
Ansible 2.2.0.0 has a bug that ignores conditionals on includes,
detailed in [ansible #17810](https://github.com/ansible/ansible/issues/17810).

In order to keep the repo working we are pinning our bootstrap to
install only 2.1.1.0 (last working release). In order to I've had to
install from pip, as 2.1.1.0 has been unpublished from EPEL, and with
that install all the dependencies that the RPM usually pulls in
automatically.